📬🔧 Discount CRUD: tratamento de erros + JSON bonitinho

+++ app/Http/Services/Admin/DiscountService.php
@@ -3,6 +3,7 @@
 namespace App\Http\Services\Admin;
 
 use App\Http\Repository\Admin\DiscountRepository;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class DiscountService{
     public function __construct(protected DiscountRepository $discountRepository)
@@ -15,7 +16,11 @@ class DiscountService{
 
     public function findDiscountById(string $id)
     {
-        return $this->discountRepository->findById($id);
+        try {
+            return $this->discountRepository->findById($id); 
+        } catch (ModelNotFoundException $e) {
+            return null;
+        }
     }
 
     public function createDiscount(array $dataValidated)
@@ -25,13 +30,25 @@ class DiscountService{
 
     public function UpdateDiscount(array $dataValidated, string $id)
     {
-        $discount  = $this->discountRepository->findById($id);
+        $discount  = $this->findDiscountById($id);
+
+        if(!$discount)
+        {
+            return null;
+        }
+
         return $discount->update($dataValidated);
     }
 
     public function deleteDiscount(string $id)
     {
-        $discount = $this->discountRepository->findById($id);
+        $discount = $this->findDiscountById($id);
+
+        if(!$discount || $discount->product != null)
+        {
+            return null;
+        }
+
         return $discount->delete();
     }
 }
